+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Speed up `AtomicThreadPool` scaling for blocking work
+
 ## [0.15.2] - 2026-08-31
 
 - Prevent enqueues from racing `AtomicThreadPool#shutdown`