+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## [Unreleased]
 
+## [0.15.4] - 2026-09-01
+
 - Keep temporary `AtomicThreadPool` workers alive for five seconds
 - Speed up `AtomicThreadPool` scaling for blocking work
 
-## [0.15.2] - 2026-08-31
+## [0.15.3] - 2026-08-31
 
 - Prevent enqueues from racing `AtomicThreadPool#shutdown`
 - Remove interrupted waiters from `AtomicConditionVariable`